refactor: update import paths for useControls to use new structure
