+++ components/controls/controls-dark-mode.tsx
@@ -3,7 +3,7 @@
 import { Moon } from 'lucide-react';
 import { useTranslations } from 'next-intl';
 
-import { useControls } from '@/lib/params/controls';
+import { useControls } from '@/lib/params/use-controls';
 import { Tooltip } from '@/components/ui/tooltip';
 import { ControlsToggle } from './controls-toggle';
